fix(H-5): prevent integer overflow in chrom_snapshot allocation
num_chrom * max_generations is computed with int multiplication; with
num_chrom=10000 and max_generations=100000 the product overflows 32-bit
signed int *before* being multiplied by sizeof(chromosome), so malloc()
receives a small (or negative) size and silently truncates the allocation.

Subsequent writes are out-of-bounds heap writes on every iteration of the
main GA loop. The bug only manifests with large user-configured parameters,
so it tends to cause silent corruption rather than crashes.

Fix: promote both operands to size_t before multiplication, range-check
the product against SIZE_MAX/sizeof(chromosome), and bail out with a clear
diagnostic if the size would overflow or either input is non-positive.

Also fixes a leftover narrowing in StatMechEngine::merge_samples that the
C-1 type-fix exposed (int multiplicities → double weight) and which was
breaking the FlexAID link target.

+++ LIB/statmech.cpp
@@ -426,7 +426,7 @@ void StatMechEngine::merge_samples(std::span<const double> energies,
     if (energies.size() != multiplicities.size())
         throw std::invalid_argument("energies and multiplicities must have same size");
     for (size_t i = 0; i < energies.size(); ++i)
-        ensemble_.push_back({energies[i], multiplicities[i]});
+        ensemble_.push_back({energies[i], static_cast<double>(multiplicities[i])});
 }
 
 std::vector<double> StatMechEngine::serialize_energies() const {
